dont expose MONDAY secret in client js bundle
