+++ apps/web/.env.production
@@ -13,7 +13,9 @@ NEXT_PUBLIC_PRODUCT_ONCO=https://onco.unensayoparami.org/
 NEXT_PUBLIC_PRODUCT_PHYSICIANS=https://app.trialtech.es/
 
 # Monday
-NEXT_PUBLIC_MONDAY_URL='https://api.monday.com/v2'
+MONDAY_URL='https://api.monday.com/v2'
+MONDAY_GROUP=grupo_nuevo17439
+MONDAY_BOARD=5210496844
 
 # Algolia
 NEXT_PUBLIC_ALGOLIA_INDEX_NAME=multi_model_search_
