+++ apps/web/components/shared/contactUs/helpers.ts
@@ -1,3 +1,6 @@
+import { ContactUsFormRequest } from './ContactUs.types';
+import dayjs from 'dayjs';
+
 export const getInitialValues = (t: Function) => ({
   name: {
     value: '',
@@ -25,3 +28,34 @@ export const getInitialValues = (t: Function) => ({
     required: true,
   },
 });
+
+export const buildQuery = ({
+  name,
+  lastname,
+  email,
+  phone,
+  query: userQuery,
+}: ContactUsFormRequest) => {
+  const date = dayjs().format('YYYY-MM-DD');
+  const fullName = `${name} ${lastname}`;
+  const item = {
+    texto8: email,
+    texto0: phone,
+    consulta: userQuery,
+    date,
+  };
+
+  const query = `
+      mutation { 
+        create_item (
+          board_id: ${process.env.MONDAY_BOARD}, 
+          group_id: "${process.env.MONDAY_GROUP}", 
+          item_name: "${fullName}", 
+          column_values: ${JSON.stringify(JSON.stringify(item))}
+          ) 
+          { id }
+        }
+    `.replace(/(\r\n|\n|\r)/gm, '');
+
+  return query;
+};
